+++ changelog.md
@@ -1,9 +1,11 @@
+* 11/09/2015 : allow to use node-maxmind plugin for geoip filter. Geoip filter can now fetch ASN.
 * 11/09/2015 : add cache on reverse dns and geoip filters
 * 25/10/2015 : Reorganize doc
 * 25/10/2015 : Update http proxy support for node > 0.10
 * 24/10/2015 : drop 0.10 support, update test for node 4
 * 24/10/2015 : use aws-sdk for aws SQS plugin
-* 04/10/2015 : publish 0.4 on NPM
+
+* 04/10/2015 : publish 0.0.4 on NPM
 * 16/05/2015 : Allow to specify dates in computed values
 * 13/05/2015 : Add basic auth for HTTP Output plugins (#100)
 * 13/05/2015 : Add websockets support (thx to @fujifish)